Make transparent contents only 8% black
The default is 15%, which may (or may not) be too dark, so that
participants get distracted too easily. However, the transparency
should not be too light, as it serves as a hint so that listeners
see that the slide isn't fully uncovered. 8% seems to be a good
compromise. Of course, this depends on the projector...