Removed too many sections
The sections did not change anything with regards to the navigation balls, as they are controlled by the frame enviroments, not the subsections (It mearly shows some of the nav balls in a slightly highlighted color if there are multiple slides in a subsection).
This plentiful use of sections and subsections however lead to empty entries being visible in the generated pdf structure. By just removing them we regain a correct content display in pdf readers.