Clip the "thank you" page big image to the bounds of the large circle.
Utilize the \clip to constrain the image to the big circle. This leads to us being able to utilize bigger rectangular images, where we do not care for the corners.
I opted for a slight border between drawn backdrop circle (radius 38mm) and the image circle (radius 37.5mm) as having the exact same size lead to there being interpolation issues on the edges when utilizing the entire circle for an image, which looked bad.